remove old ProblemView
